refactor: makes notification torrente related
